+++ AlterEgo.lua
@@ -1,4 +1,5 @@
-AlterEgo = LibStub("AceAddon-3.0"):NewAddon("AlterEgo", "AceConsole-3.0", "AceTimer-3.0", "AceEvent-3.0", "AceBucket-3.0")
+AlterEgo = LibStub("AceAddon-3.0"):NewAddon("AlterEgo", "AceConsole-3.0", "AceTimer-3.0", "AceEvent-3.0",
+                                            "AceBucket-3.0")
 AlterEgo.Libs = {}
 AlterEgo.Libs.AceDB = LibStub:GetLibrary("AceDB-3.0")
 AlterEgo.Libs.LDB = LibStub:GetLibrary("LibDataBroker-1.1")
@@ -31,19 +32,40 @@ AlterEgo.constants = {
       collapsedWidth = 30
     }
   },
-  sortingOptions = {
-    {value = "lastUpdate",  text = "Recently played"},
-    {value = "name.asc",    text = "Name (A-Z)"},
-    {value = "name.desc",   text = "Name (Z-A)"},
-    {value = "realm.asc",   text = "Realm (A-Z)"},
-    {value = "realm.desc",  text = "Realm (Z-A)"},
-    {value = "rating.asc",  text = "Rating (Lowest)"},
-    {value = "rating.desc", text = "Rating (Highest)"},
-    {value = "ilvl.asc",    text = "Item Level (Lowest)"},
-    {value = "ilvl.desc",   text = "Item Level (Highest)"},
-    {value = "class.asc",   text = "Class (A-Z)"},
-    {value = "class.desc",  text = "Class (Z-A)"},
-  }
+  sortingOptions = {{
+    value = "lastUpdate",
+    text = "Recently played"
+  }, {
+    value = "name.asc",
+    text = "Name (A-Z)"
+  }, {
+    value = "name.desc",
+    text = "Name (Z-A)"
+  }, {
+    value = "realm.asc",
+    text = "Realm (A-Z)"
+  }, {
+    value = "realm.desc",
+    text = "Realm (Z-A)"
+  }, {
+    value = "rating.asc",
+    text = "Rating (Lowest)"
+  }, {
+    value = "rating.desc",
+    text = "Rating (Highest)"
+  }, {
+    value = "ilvl.asc",
+    text = "Item Level (Lowest)"
+  }, {
+    value = "ilvl.desc",
+    text = "Item Level (Highest)"
+  }, {
+    value = "class.asc",
+    text = "Class (A-Z)"
+  }, {
+    value = "class.desc",
+    text = "Class (Z-A)"
+  }}
 }
 function AlterEgo:OnInitialize()
   _G["BINDING_NAME_ALTEREGO"] = "Show/Hide the window"
@@ -61,7 +83,8 @@ function AlterEgo:OnInitialize()
     end,
     OnTooltipShow = function(tooltip)
       tooltip:SetText("AlterEgo", 1, 1, 1)
-      tooltip:AddLine("Click to show the character summary.", NORMAL_FONT_COLOR.r, NORMAL_FONT_COLOR.g, NORMAL_FONT_COLOR.b)
+      tooltip:AddLine("Click to show the character summary.", NORMAL_FONT_COLOR.r, NORMAL_FONT_COLOR.g,
+                      NORMAL_FONT_COLOR.b)
       local dragText = "Drag to move this icon"
       if self.db.global.minimap.lock then
         dragText = dragText .. " |cffff0000(locked)|r"
@@ -93,10 +116,11 @@ function AlterEgo:OnMythicPlusData()
     return self:ScheduleTimer("OnMythicPlusData", 1)
   end
 
-  self:RegisterBucketEvent({"BAG_UPDATE_DELAYED", "PLAYER_EQUIPMENT_CHANGED", "UNIT_INVENTORY_CHANGED", "ITEM_CHANGED"}, 3, function()
-    self:UpdateCharacterInfo()
-    self:UpdateKeystoneItem()
-  end)
+  self:RegisterBucketEvent(
+    {"BAG_UPDATE_DELAYED", "PLAYER_EQUIPMENT_CHANGED", "UNIT_INVENTORY_CHANGED", "ITEM_CHANGED"}, 3, function()
+      self:UpdateCharacterInfo()
+      self:UpdateKeystoneItem()
+    end)
   self:RegisterBucketEvent({"RAID_INSTANCE_WELCOME", "LFG_LOCK_INFO_RECEIVED", "BOSS_KILL"}, 2, RequestRaidInfo)
   self:RegisterEvent("ENCOUNTER_END", "OnEncounterEnd")
   self:RegisterEvent("MYTHIC_PLUS_CURRENT_AFFIX_UPDATE", function()
@@ -108,15 +132,19 @@ function AlterEgo:OnMythicPlusData()
   self:RegisterBucketEvent({"UPDATE_INSTANCE_INFO", "LFG_UPDATE_RANDOM_INFO"}, 3, function()
     self:UpdateRaidInstances()
   end)
-  self:RegisterBucketEvent({"CHALLENGE_MODE_COMPLETED", "CHALLENGE_MODE_RESET", "CHALLENGE_MODE_MAPS_UPDATE", "MYTHIC_PLUS_NEW_WEEKLY_RECORD"}, 3, function()
-    self:UpdateMythicPlus()
-    self:UpdateKeystoneItem()
-  end)
+  self:RegisterBucketEvent({"CHALLENGE_MODE_COMPLETED", "CHALLENGE_MODE_RESET", "CHALLENGE_MODE_MAPS_UPDATE",
+                             "MYTHIC_PLUS_NEW_WEEKLY_RECORD"}, 3, function()
+                             self:UpdateMythicPlus()
+                             self:UpdateKeystoneItem()
+                           end)
   self:RegisterEvent("PLAYER_LEVEL_UP", "UpdateDB")
   self:RegisterEvent("CHAT_MSG_SYSTEM", "OnChatMessageSystem")
-  self:RegisterBucketEvent({"BONUS_ROLL_RESULT", "QUEST_CURRENCY_LOOT_RECEIVED", "POST_MATCH_CURRENCY_REWARD_UPDATE", "PLAYER_TRADE_CURRENCY", "TRADE_CURRENCY_CHANGED", "TRADE_SKILL_CURRENCY_REWARD_RESULT", "SPELL_CONFIRMATION_PROMPT", "CHAT_MSG_CURRENCY", "CURRENCY_DISPLAY_UPDATE"}, 3, function()
-    self:UpdateCharacterInfo()
-  end)
+  self:RegisterBucketEvent({"BONUS_ROLL_RESULT", "QUEST_CURRENCY_LOOT_RECEIVED", "POST_MATCH_CURRENCY_REWARD_UPDATE",
+                             "PLAYER_TRADE_CURRENCY", "TRADE_CURRENCY_CHANGED", "TRADE_SKILL_CURRENCY_REWARD_RESULT",
+                             "SPELL_CONFIRMATION_PROMPT", "CHAT_MSG_CURRENCY", "CURRENCY_DISPLAY_UPDATE"}, 3,
+                           function()
+                             self:UpdateCharacterInfo()
+                           end)
 
   self:loadGameData()
   self:MigrateDB()
@@ -137,7 +165,8 @@ function AlterEgo:OnChatMessageSystem(_, msg)
   if not groupChannel or not self.db.global.announceResets or IsInInstance() or not UnitIsGroupLeader("player") then
     return
   end
-  local resetPatterns = {INSTANCE_RESET_SUCCESS, INSTANCE_RESET_FAILED, INSTANCE_RESET_FAILED_OFFLINE, INSTANCE_RESET_FAILED_ZONING}
+  local resetPatterns = {INSTANCE_RESET_SUCCESS, INSTANCE_RESET_FAILED, INSTANCE_RESET_FAILED_OFFLINE,
+    INSTANCE_RESET_FAILED_ZONING}
   AE_table_foreach(resetPatterns, function(resetPattern)
     if msg:match("^" .. resetPattern:gsub("%%s", ".+") .. "$") then
       SendChatMessage(self.constants.prefix .. msg, groupChannel)
@@ -145,59 +174,61 @@ function AlterEgo:OnChatMessageSystem(_, msg)
   end)
 end
 
-function AlterEgo:AnnounceKeystones(chatType, multiline)
+function AlterEgo:AnnounceKeystones(chatType)
   local characters = self:GetCharacters()
   local dungeons = self:GetDungeons()
-  local msg = "My Keystones: "
+  local multiline = self.db.global.announceKeystones.multiline
+  local multilineNames = self.db.global.announceKeystones.multilineNames
+  local keystones = {}
+  local keystonesCompact = {}
 
   if AE_table_count(characters) < 1 then
     self:Print("No announcement: You have no characters saved.")
     return
   end
 
-  local keystoneCharacters = AE_table_filter(characters, function(character)
+  AE_table_foreach(characters, function(character)
+    local keystone = character.mythicplus.keystone
     local dungeon
-    if type(character.mythicplus.keystone.challengeModeID) == "number" and character.mythicplus.keystone.challengeModeID > 0 then
-      dungeon = AE_table_get(dungeons, "challengeModeID", character.mythicplus.keystone.challengeModeID)
-    elseif type(character.mythicplus.keystone.mapId) == "number" and character.mythicplus.keystone.mapId > 0 then
-      dungeon = AE_table_get(dungeons, "mapId", character.mythicplus.keystone.mapId)
+
+    if type(keystone.level) ~= "number" or keystone.level < 1 then
+      return
     end
-    if dungeon and type(character.mythicplus.keystone.level) == "number" and character.mythicplus.keystone.level > 0 then
-      return true
+
+    if type(keystone.challengeModeID) == "number" and keystone.challengeModeID > 0 then
+      dungeon = AE_table_get(dungeons, "challengeModeID", keystone.challengeModeID)
+    elseif type(keystone.mapId) == "number" and keystone.mapId > 0 then
+      dungeon = AE_table_get(dungeons, "mapId", keystone.mapId)
     end
-    return false
+
+    if not dungeon then
+      return
+    end
+
+    local text = dungeon.abbr .. " +" .. tostring(keystone.level)
+    table.insert(keystones, {
+      text = text,
+      itemLink = keystone.itemLink,
+      characterName = character.info.name
+    })
+    table.insert(keystonesCompact, text)
   end)
 
-  if AE_table_count(keystoneCharacters) < 1 then
+  if AE_table_count(keystones) < 1 then
     self:Print("No announcement: You have no keystones saved.")
     return
   end
 
-  local count = 0
-  AE_table_foreach(keystoneCharacters, function(character)
-    local dungeon
-    if type(character.mythicplus.keystone.challengeModeID) == "number" and character.mythicplus.keystone.challengeModeID > 0 then
-      dungeon = AE_table_get(dungeons, "challengeModeID", character.mythicplus.keystone.challengeModeID)
-    elseif type(character.mythicplus.keystone.mapId) == "number" and character.mythicplus.keystone.mapId > 0 then
-      dungeon = AE_table_get(dungeons, "mapId", character.mythicplus.keystone.mapId)
-    end
-    if dungeon then
-      if type(character.mythicplus.keystone.level) == "number" and character.mythicplus.keystone.level > 0 then
-        local currentKeystone = dungeon.abbr .. " +" .. tostring(character.mythicplus.keystone.level)
-        if multiline then
-          SendChatMessage(character.info.name .. ": " .. (character.mythicplus.keystone.itemLink and character.mythicplus.keystone.itemLink or currentKeystone), chatType)
-        else
-          if count > 0 then
-            msg = msg .. " || "
-          end
-          msg = msg .. currentKeystone
-        end
-        count = count + 1
+  if multiline then
+    AE_table_foreach(keystones, function(keystone)
+      local chatMessage = keystone.itemLink and keystone.itemLink or keystone.text
+      if multilineNames == true then
+        chatMessage = keystone.characterName .. ": " .. chatMessage
       end
-    end
-  end)
-
-  if not multiline then
-    SendChatMessage(self.constants.prefix .. msg, chatType)
+      SendChatMessage(chatMessage, chatType)
+    end)
+    return
   end
+
+  SendChatMessage(self.constants.prefix .. "My keystones: " .. table.concat(keystonesCompact, " || "), chatType)
 end